[WebNN EP] Automatically move input CPU tensors to ml-tensor (#23073)
### Description
If it would improve performance, this patch moves the CPU to ml-tensor
before sending the to the ONNXRuntime WebNN EP.

### Motivation and Context
We are currently performing 2 extra copies on input tensors located in
the CPU when using the WebNN EP (JS -(copy)-> wasm heap -(copy)-> JS ->
WebNN API). This patch removes these extra copies.